Fixed windows platform x64 compatibility issues.

+++ code/win32/win_main.c
@@ -290,7 +290,7 @@ char **Sys_ListFiles( const char *directory, const char *extension, char *filter
 	char		**listCopy;
 	char		*list[MAX_FOUND_FILES];
 	struct _finddata_t findinfo;
-	int			findhandle;
+	intptr_t  findhandle;
 	int			flag;
 	int			i;
